fastapi_app/requirements.txt
httpx==0.24.1
fastapi==0.100.0
uvicorn==0.22.0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-instrumentation-logging==0.39b0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
